test(config): consolidate config resolution tests and add env injection
- Merge cli-config-resolve.spec.ts and mcp-config-resolve.spec.ts into
  config-resolve.spec.ts, deduplicating shared browser/channel, sandbox,
  timeout, viewport, and merge-order tests.
- Add optional `env` parameter to configFromEnv / resolveCLIConfigForMCP /
  resolveCLIConfigForCLI so tests pass a clean env instead of
  saving/restoring process.env.
- Move sandbox and cdpHeaders unit tests out of config.spec.ts into the
  new file; config.spec.ts now only has integration tests.
